ref: code structure for improved readability and maintainability

+++ FA25_PRN222_3W1_ASM5_SE170488_TienPVK/LionPetManagement_PhanVuKhanhTien/Pages/LionProfiles/Delete.cshtml.cs
@@ -49,8 +49,12 @@ namespace LionPetManagement_PhanVuKhanhTien.Pages.LionProfiles
             {
                 return NotFound();
             }
-
-            await _service.DeleteAsync(id.Value);
+            var obj = await _service.GetByIdAsync(id.Value);
+            if (obj != null)
+            {
+                LionProfile = obj;
+                await _service.DeleteAsync(id.Value);
+            }
 
             return RedirectToPage("./Index");
         }
